+++ babel.config.js
@@ -1,0 +1,19 @@
+module.exports = {
+  env: {
+    build: {
+      presets: ['@babel/preset-env'],
+    },
+    test: {
+      presets: [
+        [
+          '@babel/preset-env',
+          {
+            targets: {
+              node: 'current',
+            },
+          },
+        ],
+      ],
+    },
+  },
+}