+++ test/__helpers__.js
@@ -5,11 +5,18 @@ import { describe, it, beforeEach, afterEach, before } from 'mocha'
 const MockXMLHttpRequest = require('mock-xmlhttprequest')
 
 export const testKinka = (kinkaPath, instanceName, beforeHook) => {
-  describe(`${instanceName} : `, () => {
+  describe(`test kinka (type: ${instanceName}, path: ${kinkaPath})`, () => {
     before(() => {
       if (beforeHook) beforeHook()
     })
-    const kinka = require(kinkaPath)
+
+    const kinka = (() => {
+      const module = require(kinkaPath)
+      return Object.keys(module).length === 1 && module.default
+        ? module.default
+        : module
+    })()
+
     it('should be defined', () => {
       expect(!!kinka).to.be.equal(true)
     })
@@ -17,12 +24,26 @@ export const testKinka = (kinkaPath, instanceName, beforeHook) => {
     const testMethod = (name, options = {}) => {
       const methodName = name.toUpperCase()
       const api = options.api || kinka
+
+      if (!api) {
+        throw new Error('api is not defined')
+      }
+
       const baseURL = options.baseURL || global.location.origin
       const path = `/test-path/method-${name}/custom-method-${Boolean(
         options.testCustom
       )}`
       const url = `${baseURL}${path}`
       const httpMock = nock(baseURL)
+
+      if (typeof api[options.testCustom ? 'custom' : name] !== 'function') {
+        throw new Error(
+          `Property with name "${
+            options.testCustom ? 'custom' : name
+          }" is not existing in api`
+        )
+      }
+
       const testFunc = options.testCustom
         ? api.custom.bind(api, name)
         : api[name].bind(api)
@@ -76,6 +97,7 @@ export const testKinka = (kinkaPath, instanceName, beforeHook) => {
 
     describe('basic methods : ', () => {
       const methods = [
+        // FIXME: property 'delete' is not existing in api
         'delete',
         'get',
         'head',
@@ -88,9 +110,12 @@ export const testKinka = (kinkaPath, instanceName, beforeHook) => {
     })
 
     describe('custom methods : ', () => {
-      const customMethods = ['permanent', 'kill', 'stop', 'move']
-      const api = kinka.create({ customMethods })
-      customMethods.forEach(name => testMethod(name, { api }))
+      describe('using "kinka[methodName] : ', () => {
+        const customMethods = ['permanent', 'kill', 'stop', 'move']
+        const api = kinka.create({ customMethods })
+        customMethods.forEach(name => testMethod(name, { api }))
+      })
+
       describe('using "kinka.custom" function : ', () => {
         const customMethods = ['foo', 'bar', 'baz', 'moveout']
         customMethods.forEach(name => testMethod(name, { testCustom: true }))